Support for integration with a parent redux store (supports multiple instances)
A redux store can now (optionally) be passed to this component by setting the "store" and "instanceKey" prop. The state for this component will then be read from `globalState.PerformrRunnerResultGraph[instanceKey]`, dispatched actions will also contain the instanceKey. The redux reducer function that is needed to handle the actions can be created by calling `createReducer(instanceKey)`

If these props are not set, a private store will be created